+++ packages/skeleton-react/src/lib/components/Navigation/Navigation.test.tsx
@@ -1,21 +1,21 @@
 import { describe, expect, it } from 'vitest';
 import { render } from '@testing-library/react';
-import { Nav } from '$lib/index.js';
+import { Navigation } from '$lib/index.js';
 
 describe('Navigation', () => {
-	describe('<Nav.Rail>', () => {
+	describe('<Navigation.Rail>', () => {
 		const testId = 'nav-rail';
 
 		it('Renders the component', () => {
-			const { getByTestId } = render(<Nav.Rail></Nav.Rail>);
+			const { getByTestId } = render(<Navigation.Rail></Navigation.Rail>);
 			expect(getByTestId(testId)).toBeInTheDocument();
 		});
 
 		it('Renders the children', () => {
 			const { getByTestId } = render(
-				<Nav.Rail>
+				<Navigation.Rail>
 					<div data-testid="child"></div>
-				</Nav.Rail>
+				</Navigation.Rail>
 			);
 			expect(getByTestId('child')).toBeInTheDocument();
 		});
@@ -23,7 +23,7 @@ describe('Navigation', () => {
 		for (const prop of ['base', 'background', 'padding', 'height', 'classes']) {
 			it(`Correctly applies the \`${prop}\` prop`, () => {
 				const value = 'foo';
-				const { getByTestId } = render(<Nav.Rail {...{ [prop]: value }} />);
+				const { getByTestId } = render(<Navigation.Rail {...{ [prop]: value }} />);
 				expect(getByTestId(testId)).toHaveClass(value);
 			});
 		}
@@ -31,10 +31,12 @@ describe('Navigation', () => {
 		it('Correctly applies the `width` or `widthExpanded` prop based on the `expanded` prop', () => {
 			const width = 'width';
 			const widthExpanded = 'widthExpanded';
-			const { getByTestId, rerender } = render(<Nav.Rail width={width} widthExpanded={widthExpanded} expanded={false}></Nav.Rail>);
+			const { getByTestId, rerender } = render(
+				<Navigation.Rail width={width} widthExpanded={widthExpanded} expanded={false}></Navigation.Rail>
+			);
 			expect(getByTestId(testId)).toHaveClass(width);
 			expect(getByTestId(testId)).not.toHaveClass(widthExpanded);
-			rerender(<Nav.Rail width={width} widthExpanded={widthExpanded} expanded={true}></Nav.Rail>);
+			rerender(<Navigation.Rail width={width} widthExpanded={widthExpanded} expanded={true}></Navigation.Rail>);
 			expect(getByTestId(testId)).not.toHaveClass(width);
 			expect(getByTestId(testId)).toHaveClass(widthExpanded);
 		});
@@ -43,7 +45,7 @@ describe('Navigation', () => {
 			const testId = 'nav-rail-header';
 
 			it('Renders the component', () => {
-				const { getByTestId } = render(<Nav.Rail header="header"></Nav.Rail>);
+				const { getByTestId } = render(<Navigation.Rail header="header"></Navigation.Rail>);
 				expect(getByTestId(testId)).toBeInTheDocument();
 				expect(getByTestId(testId)).toHaveTextContent('header');
 			});
@@ -51,7 +53,7 @@ describe('Navigation', () => {
 			for (const prop of ['headerBase', 'headerFlexDirection', 'headerJustify', 'headerItems', 'headerGap', 'headerClasses']) {
 				it(`Correctly applies the \`${prop}\` prop`, () => {
 					const value = 'foo';
-					const { getByTestId } = render(<Nav.Rail header="header" {...{ [prop]: value }}></Nav.Rail>);
+					const { getByTestId } = render(<Navigation.Rail header="header" {...{ [prop]: value }}></Navigation.Rail>);
 					expect(getByTestId(testId)).toHaveClass(value);
 				});
 			}
@@ -61,7 +63,7 @@ describe('Navigation', () => {
 			const testId = 'nav-rail-tiles';
 
 			it('Renders the component', () => {
-				const { getByTestId } = render(<Nav.Rail>tiles</Nav.Rail>);
+				const { getByTestId } = render(<Navigation.Rail>tiles</Navigation.Rail>);
 				expect(getByTestId(testId)).toBeInTheDocument();
 				expect(getByTestId(testId)).toHaveTextContent('tiles');
 			});
@@ -69,7 +71,7 @@ describe('Navigation', () => {
 			for (const prop of ['tilesBase', 'tilesFlexDirection', 'tilesJustify', 'tilesItems', 'tilesGap', 'tilesClasses']) {
 				it(`Correctly applies the \`${prop}\` prop`, () => {
 					const value = 'foo';
-					const { getByTestId } = render(<Nav.Rail {...{ [prop]: value }}>tiles</Nav.Rail>);
+					const { getByTestId } = render(<Navigation.Rail {...{ [prop]: value }}>tiles</Navigation.Rail>);
 					expect(getByTestId(testId)).toHaveClass(value);
 				});
 			}
@@ -79,7 +81,7 @@ describe('Navigation', () => {
 			const testId = 'nav-rail-footer';
 
 			it('Renders the component', () => {
-				const { getByTestId } = render(<Nav.Rail footer="footer"></Nav.Rail>);
+				const { getByTestId } = render(<Navigation.Rail footer="footer"></Navigation.Rail>);
 				expect(getByTestId(testId)).toBeInTheDocument();
 				expect(getByTestId(testId)).toHaveTextContent('footer');
 			});
@@ -87,26 +89,26 @@ describe('Navigation', () => {
 			for (const prop of ['footerBase', 'footerFlexDirection', 'footerJustify', 'footerItems', 'footerGap', 'footerClasses']) {
 				it(`Correctly applies the \`${prop}\` prop`, () => {
 					const value = 'foo';
-					const { getByTestId } = render(<Nav.Rail footer="footer" {...{ [prop]: value }} />);
+					const { getByTestId } = render(<Navigation.Rail footer="footer" {...{ [prop]: value }} />);
 					expect(getByTestId(testId)).toHaveClass(value);
 				});
 			}
 		});
 	});
 
-	describe('<Nav.Bar>', () => {
+	describe('<Navigation.Bar>', () => {
 		const testId = 'nav-bar';
 
 		it('Renders the component', () => {
-			const { getByTestId } = render(<Nav.Bar></Nav.Bar>);
+			const { getByTestId } = render(<Navigation.Bar></Navigation.Bar>);
 			expect(getByTestId(testId)).toBeInTheDocument();
 		});
 
 		it('Renders the children', () => {
 			const { getByTestId } = render(
-				<Nav.Bar>
+				<Navigation.Bar>
 					<div data-testid="child"></div>
-				</Nav.Bar>
+				</Navigation.Bar>
 			);
 			expect(getByTestId('child')).toBeInTheDocument();
 		});
@@ -114,31 +116,31 @@ describe('Navigation', () => {
 		for (const prop of ['base', 'background', 'width', 'height', 'padding', 'classes']) {
 			it(`Correctly applies the \`${prop}\` prop`, () => {
 				const value = 'foo';
-				const { getByTestId } = render(<Nav.Bar {...{ [prop]: value }} />);
+				const { getByTestId } = render(<Navigation.Bar {...{ [prop]: value }} />);
 				expect(getByTestId(testId)).toHaveClass(value);
 			});
 		}
 	});
 
-	describe('<Nav.Tile>', () => {
+	describe('<Navigation.Tile>', () => {
 		const testId = 'nav-tile';
 
 		it('Renders the component', () => {
 			const { getByTestId } = render(
-				<Nav.Rail>
-					<Nav.Tile></Nav.Tile>
-				</Nav.Rail>
+				<Navigation.Rail>
+					<Navigation.Tile></Navigation.Tile>
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).toBeInTheDocument();
 		});
 
 		it('Renders the children', () => {
 			const { getByTestId } = render(
-				<Nav.Rail>
-					<Nav.Tile>
+				<Navigation.Rail>
+					<Navigation.Tile>
 						<div data-testid="child"></div>
-					</Nav.Tile>
-				</Nav.Rail>
+					</Navigation.Tile>
+				</Navigation.Rail>
 			);
 			expect(getByTestId('child')).toBeInTheDocument();
 		});
@@ -146,7 +148,7 @@ describe('Navigation', () => {
 		for (const prop of ['base', 'width', 'aspect', 'background', 'hover', 'gap', 'rounded']) {
 			it(`Correctly applies the \`${prop}\` prop`, () => {
 				const value = 'foo';
-				const { getByTestId } = render(<Nav.Tile {...{ [prop]: value }} />);
+				const { getByTestId } = render(<Navigation.Tile {...{ [prop]: value }} />);
 				expect(getByTestId(testId)).toHaveClass(value);
 			});
 		}
@@ -155,16 +157,16 @@ describe('Navigation', () => {
 			const padding = 'padding';
 			const expandedPadding = 'expandedPadding';
 			const { getByTestId, rerender } = render(
-				<Nav.Rail expanded={false}>
-					<Nav.Tile padding={padding} expandedPadding={expandedPadding}></Nav.Tile>);
-				</Nav.Rail>
+				<Navigation.Rail expanded={false}>
+					<Navigation.Tile padding={padding} expandedPadding={expandedPadding}></Navigation.Tile>);
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).toHaveClass(padding);
 			expect(getByTestId(testId)).not.toHaveClass(expandedPadding);
 			rerender(
-				<Nav.Rail expanded={true}>
-					<Nav.Tile padding={padding} expandedPadding={expandedPadding}></Nav.Tile>);
-				</Nav.Rail>
+				<Navigation.Rail expanded={true}>
+					<Navigation.Tile padding={padding} expandedPadding={expandedPadding}></Navigation.Tile>);
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).not.toHaveClass(padding);
 			expect(getByTestId(testId)).toHaveClass(expandedPadding);
@@ -174,16 +176,16 @@ describe('Navigation', () => {
 			const gap = 'gap';
 			const expandedGap = 'expandedGap';
 			const { getByTestId, rerender } = render(
-				<Nav.Rail expanded={false}>
-					<Nav.Tile gap={gap} expandedGap={expandedGap}></Nav.Tile>);
-				</Nav.Rail>
+				<Navigation.Rail expanded={false}>
+					<Navigation.Tile gap={gap} expandedGap={expandedGap}></Navigation.Tile>);
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).toHaveClass(gap);
 			expect(getByTestId(testId)).not.toHaveClass(expandedGap);
 			rerender(
-				<Nav.Rail expanded={true}>
-					<Nav.Tile gap={gap} expandedGap={expandedGap}></Nav.Tile>);
-				</Nav.Rail>
+				<Navigation.Rail expanded={true}>
+					<Navigation.Tile gap={gap} expandedGap={expandedGap}></Navigation.Tile>);
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).not.toHaveClass(gap);
 			expect(getByTestId(testId)).toHaveClass(expandedGap);
@@ -193,16 +195,16 @@ describe('Navigation', () => {
 			const classes = 'classes';
 			const expandedClasses = 'expandedClasses';
 			const { getByTestId, rerender } = render(
-				<Nav.Rail expanded={false}>
-					<Nav.Tile classes={classes} expandedClasses={expandedClasses}></Nav.Tile>);
-				</Nav.Rail>
+				<Navigation.Rail expanded={false}>
+					<Navigation.Tile classes={classes} expandedClasses={expandedClasses}></Navigation.Tile>);
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).toHaveClass(classes);
 			expect(getByTestId(testId)).not.toHaveClass(expandedClasses);
 			rerender(
-				<Nav.Rail expanded={true}>
-					<Nav.Tile classes={classes} expandedClasses={expandedClasses}></Nav.Tile>);
-				</Nav.Rail>
+				<Navigation.Rail expanded={true}>
+					<Navigation.Tile classes={classes} expandedClasses={expandedClasses}></Navigation.Tile>);
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).not.toHaveClass(classes);
 			expect(getByTestId(testId)).toHaveClass(expandedClasses);
@@ -211,15 +213,15 @@ describe('Navigation', () => {
 		it('Correctly applies the `active` prop based on wether the tile is active', async () => {
 			const active = 'active';
 			const { getByTestId, rerender } = render(
-				<Nav.Rail>
-					<Nav.Tile active={active} selected={false}></Nav.Tile>
-				</Nav.Rail>
+				<Navigation.Rail>
+					<Navigation.Tile active={active} selected={false}></Navigation.Tile>
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).not.toHaveClass('active');
 			rerender(
-				<Nav.Rail>
-					<Nav.Tile active={active} selected={true}></Nav.Tile>
-				</Nav.Rail>
+				<Navigation.Rail>
+					<Navigation.Tile active={active} selected={true}></Navigation.Tile>
+				</Navigation.Rail>
 			);
 			expect(getByTestId(testId)).toHaveClass('active');
 		});
@@ -229,9 +231,9 @@ describe('Navigation', () => {
 
 			it('Renders the component', () => {
 				const { getByTestId } = render(
-					<Nav.Rail>
-						<Nav.Tile label="label"></Nav.Tile>
-					</Nav.Rail>
+					<Navigation.Rail>
+						<Navigation.Tile label="label"></Navigation.Tile>
+					</Navigation.Rail>
 				);
 				expect(getByTestId(testId)).toBeInTheDocument();
 				expect(getByTestId(testId)).toHaveTextContent('label');
@@ -241,9 +243,9 @@ describe('Navigation', () => {
 				it(`Correctly applies the \`${prop}\` prop`, () => {
 					const value = 'foo';
 					const { getByTestId } = render(
-						<Nav.Rail>
-							<Nav.Tile label="label" {...{ [prop]: value }}></Nav.Tile>
-						</Nav.Rail>
+						<Navigation.Rail>
+							<Navigation.Tile label="label" {...{ [prop]: value }}></Navigation.Tile>
+						</Navigation.Rail>
 					);
 					expect(getByTestId(testId)).toHaveClass(value);
 				});
@@ -255,9 +257,9 @@ describe('Navigation', () => {
 
 			it('Renders the component', () => {
 				const { getByTestId } = render(
-					<Nav.Rail expanded={true}>
-						<Nav.Tile labelExpanded="labelExpanded"></Nav.Tile>
-					</Nav.Rail>
+					<Navigation.Rail expanded={true}>
+						<Navigation.Tile labelExpanded="labelExpanded"></Navigation.Tile>
+					</Navigation.Rail>
 				);
 				expect(getByTestId(testId)).toBeInTheDocument();
 				expect(getByTestId(testId)).toHaveTextContent('labelExpanded');
@@ -267,9 +269,9 @@ describe('Navigation', () => {
 				it(`Correctly applies the \`${prop}\` prop`, () => {
 					const value = 'foo';
 					const { getByTestId } = render(
-						<Nav.Rail expanded={true}>
-							<Nav.Tile labelExpanded="labelExpanded" {...{ [prop]: value }}></Nav.Tile>
-						</Nav.Rail>
+						<Navigation.Rail expanded={true}>
+							<Navigation.Tile labelExpanded="labelExpanded" {...{ [prop]: value }}></Navigation.Tile>
+						</Navigation.Rail>
 					);
 					expect(getByTestId(testId)).toHaveClass(value);
 				});