+++ packages/skeleton-react/src/lib/components/Navigation/Navigation.tsx
@@ -157,6 +157,7 @@ export const NavTile: React.FC<NavTileProps> = ({
 	labelExpanded,
 	title,
 	selected = false,
+	type = 'button',
 	// Root
 	base = 'flex items-center',
 	width = 'w-full',
@@ -188,7 +189,6 @@ export const NavTile: React.FC<NavTileProps> = ({
 
 	// Local
 	const TileElement = href ? 'a' : 'button';
-	const type = href ? undefined : 'button';
 	const role = href ? undefined : 'button';
 
 	// Reactive
@@ -209,7 +209,7 @@ export const NavTile: React.FC<NavTileProps> = ({
 			className={`${base} ${width} ${rounded} ${rxBackground} ${rxMode}`}
 			href={href}
 			target={target}
-			type={type}
+			type={TileElement === 'button' ? type : undefined}
 			title={title}
 			role={role}
 			onClick={onClickHandler}
@@ -237,7 +237,7 @@ export const NavTile: React.FC<NavTileProps> = ({
 
 // Exports ---
 
-export const Nav = Object.assign(() => null, {
+export const Navigation = Object.assign(() => null, {
 	Rail: NavRail,
 	Bar: NavBar,
 	Tile: NavTile