+++ sites/next.skeleton.dev/src/examples/components/navigation/Example.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Nav } from '@skeletonlabs/skeleton-react';
+import { Navigation } from '@skeletonlabs/skeleton-react';
 import {
 	Menu as IconMenu,
 	Folder as IconFolder,
@@ -14,15 +14,15 @@ export const Page: React.FC = () => {
 		<div className="card border-surface-100-900 grid h-[640px] w-full grid-cols-[auto_1fr] border-[1px]">
 			{/* Component */}
 			{/* prettier-ignore */}
-			<Nav.Rail
-				header={<Nav.Tile href="/" title="Menu"><IconMenu /></Nav.Tile>}
-				footer={<Nav.Tile href="/settings" title="settings"><IconSettings /></Nav.Tile>}
+			<Navigation.Rail
+				header={<Navigation.Tile href="/" title="Menu"><IconMenu /></Navigation.Tile>}
+				footer={<Navigation.Tile href="/settings" title="settings"><IconSettings /></Navigation.Tile>}
 			>
-				<Nav.Tile id="0" label="Files" href="#"><IconFolder /></Nav.Tile>
-				<Nav.Tile id="1" label="Images" href="#"><IconImage /></Nav.Tile>
-				<Nav.Tile id="2" label="Music" href="#"><IconMusic /></Nav.Tile>
-				<Nav.Tile id="3" label="Videos" href="#"><IconVideo /></Nav.Tile>
-			</Nav.Rail>
+				<Navigation.Tile id="0" label="Files" href="#"><IconFolder /></Navigation.Tile>
+				<Navigation.Tile id="1" label="Images" href="#"><IconImage /></Navigation.Tile>
+				<Navigation.Tile id="2" label="Music" href="#"><IconMusic /></Navigation.Tile>
+				<Navigation.Tile id="3" label="Videos" href="#"><IconVideo /></Navigation.Tile>
+			</Navigation.Rail>
 			{/* Content */}
 			<div className="flex items-center justify-center">
 				<p className="opacity-20">(Content)</p>